Fix warnings in cocos2dx target of XCode project.

+++ cocos/base/CCConsole.cpp
@@ -638,7 +638,7 @@ void Console::commandTouch(int fd, const std::string& args)
                 float x = std::atof(argv[1].c_str());
                 float y = std::atof(argv[2].c_str());
 
-                srand (time(NULL));
+                srand ((unsigned)time(NULL));
                 _touchId = rand();
                 Scheduler *sched = Director::getInstance()->getScheduler();
                 sched->performFunctionInCocosThread( [&](){
@@ -666,7 +666,7 @@ void Console::commandTouch(int fd, const std::string& args)
                 float x2 = std::atof(argv[3].c_str());
                 float y2 = std::atof(argv[4].c_str());
 
-                srand (time(NULL));
+                srand ((unsigned)time(NULL));
                 _touchId = rand();
 
                 Scheduler *sched = Director::getInstance()->getScheduler();
@@ -819,7 +819,7 @@ void Console::commandUpload(int fd)
     fclose(fp);
 }
 
-ssize_t Console::readBytes(int fd, char* buffer, int maxlen, bool* more)
+ssize_t Console::readBytes(int fd, char* buffer, size_t maxlen, bool* more)
 {
     ssize_t n, rc;
     char c, *ptr = buffer;
@@ -929,7 +929,7 @@ bool Console::parseCommand(int fd)
 //
 
 
-ssize_t Console::readline(int fd, char* ptr, int maxlen)
+ssize_t Console::readline(int fd, char* ptr, size_t maxlen)
 {
     ssize_t n, rc;
     char c;
